fix(api): consolidate duplicate guards and fix integration decryption
- Consolidate JwtAuthGuard, SdkKeyGuard, TenantGuard, RolesGuard to common/guards/
- Update auth/guards/ and modules/auth/guards/ to re-export from common/guards
- Add RolesGuard to common/guards (previously only in auth/guards)
- Update all controller imports to use common/guards
- Add try/catch around integration config decryption in worker
- Auto-disable integrations with corrupted credentials
- Log decryption errors without crashing worker

Closes #131
