+++ apps/api/src/common/guards/index.ts
@@ -1,3 +1,5 @@
 export * from './jwt-auth.guard';
 export * from './sdk-key.guard';
 export * from './tenant.guard';
+export * from './roles.guard';
+export * from './ip-whitelist.guard';
